Final version

+++ code2048/v4_2048_32.c
@@ -5,7 +5,7 @@
 #include<time.h>
 
 int N = 2048;
-int A[2048][2048];
+float A[2048][2048];
 int array1[2048][2048];
 int array2[2048][2048];
 int B = 512;
